requirements.txt
black==22.10.0
Django==4.1.2
django-debug-toolbar==3.7.0
django-split-settings==1.2.0
dynaconf==3.1.11
flake8==5.0.4
psutil==5.9.4
psycopg2-binary==2.9.4
py==1.11.0
pycodestyle==2.9.1
pydantic==1.10.2
python-dotenv==0.21.0
pyflakes==2.5.0
pytest==7.1.3
python-gettext==4.1
sqlparse==0.4.3
uwsgi==2.0.20